test correction and structuration
