+++ requirements.txt
@@ -3,3 +3,4 @@ mir_eval
 pretty_midi
 matplotlib
 setuptools
+tabulate
